requirements.txt
click==8.0.3
colorama==0.4.5
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dotenv==0.20.0
Werkzeug==2.0.2
